feat: migrate products pages to App Router (Task 2.5)
