+++ src/app/products/list/page.js
@@ -1,0 +1,5 @@
+import { redirect } from "next/navigation";
+
+export default function ProductsListIndexPage() {
+  redirect("/products/list/1");
+}
